+++ lib/flutter_sync_tree.dart
@@ -5,3 +5,5 @@ export 'src/sync_models.dart';
 export 'src/sync_node.dart';
 export 'src/sync_leaf.dart';
 export 'src/sync_composite.dart';
+export 'src/sync_types.dart';
+export 'src/sync_state.dart';